Create Hamiltonian Cycle
Determine whether a given undirected graph contains the Hamiltonian Cycle or not. Hamiltonian Cycle in an undirected graph is a path that visits each vertex exactly once. A Hamiltonian cycle is a Hamiltonian Path such that there is an edge from the last vertex to the first vertex of the Hamiltonian Path.