Create Minimum Cost
You are given a grid of size N x M represented by a 2D array cost[N][M]. Each cell in the grid contains a non-negative integer representing the cost associated with that cell. Your task is to find the minimum cost to travel from the top-left cell (0,0) to the bottom-right cell (N-1,M-1) while adhering to the following rules:

Rules:
You can only move right or down from your current cell. Each step you take incurs a cost equal to the value of the cell you are moving to.